Use a small stack allocated buffer instead of malloc/free in 'doBuiltinPrintf' and 'doBuiltinScanf'.
Because 'doBuiltinPrintf' runs for every "%X" in the formatted string of the builtin 'printf', using malloc/free is unnecessary overhead.
This overhead can be avoided when the formatted string is small, which it usually is.
There is the question of how big the small buffer should be. For now it's 2048.
Unfortunately gcc now gives us bogus? warnings: " warning: 'strncpy' destination unchanged after copying no bytes [-Wstringop-truncation]".